requirements-tests.txt
-r requirements.txt
ipdb
pytest-cov
pillow
tox
wand>=0.3
